Tc2 11

See merge request foxminded-teamconnect-2/tracker!12

+++ src/main/java/com/ua/teamconnect/tracker/repository/UserPositionRepository.java
@@ -17,8 +17,8 @@ public interface UserPositionRepository extends CrudRepository<UserPosition, Use
     select up from UserPosition up join fetch up.position p join fetch p.department
         where up.id.userId=:userId and up.startDate <= :now and (up.endDate is null or :now < up.endDate)
     """)
-    List<UserPosition> findByUserIdAndNow(Long userId, LocalDate now);
+    List<UserPosition> findByUserIdAndNow(Integer userId, LocalDate now);
 
     @Query("select min(up.startDate) hire_date from UserPosition up where up.id.userId=:userId")
-    Optional<LocalDate> findHireDateByUserId(Long userId);
+    Optional<LocalDate> findHireDateByUserId(Integer userId);
 }
